+++ client/src/style/css/LabelsInputs.css
@@ -1,0 +1,35 @@
+@font-face {
+    font-family: fonteLabel;
+    src: url(../font/SuisseIntlCond-Semibold-WebS.woff);
+}
+
+.labelInput label{
+    font-family: fonteLabel;
+    font-weight: 550;
+    font-size: 16px;
+    display: flex;
+    color: #3E3C39;
+
+    margin-left: 7vh;
+    height: 3.5vh;
+    
+}
+
+.labelInput input{
+    width: 30vh;
+    height: 4vh;
+    margin-left: 12.5%;
+    padding-left: 2.5vh;
+    font-size: 12px;
+    display: flex;
+    letter-spacing: 0.01em;
+    background: #FFFFFF;
+
+    box-shadow: 1px 4px 15px rgba(0, 0, 0, 0.08);
+    border-radius: 30px;
+    border: #ffffff;
+    
+    color: #AEADAA;
+
+}
+
